refactoring code contacts
