+++ src/layout/sections/skills/Skills.tsx
@@ -5,22 +5,7 @@ import image from '../../../assets/images/skills.webp'
 import {SectionTitle} from '../../../components/SectionTitle';
 import {Container} from '../../../components/Container';
 import {S} from './Skills_Styles'
-
-
-const SkillData: SkillPropsType[] = [
-    {iconId: 'typeScript', titleId: 'Type Script'},
-    {iconId: 'javaScript', titleId: 'Java Script'},
-    {iconId: 'react', titleId: 'React', viewBox: '0 0 40 38'},
-    {iconId: 'redux', titleId: 'Redux'},
-    {iconId: 'restAPI', titleId: 'Rest API'},
-    {iconId: 'git', titleId: 'Git'},
-    {iconId: 'html', titleId: 'HTML'},
-    {iconId: 'css', titleId: 'CSS'},
-    {iconId: 'sass', titleId: 'SASS'},
-    {iconId: 'bootstrap', titleId: 'Bootstrap'},
-    {iconId: 'styledComponents', titleId: 'Styled Components', viewBox: '0 0 32 32'},
-    {iconId: 'materialUI', titleId: 'Material UI', viewBox: '0 0 256 204'},
-]
+import {Data} from './../../../data/Data'
 
 export const Skills = () => {
     return (
@@ -31,7 +16,7 @@ export const Skills = () => {
                     <S.SkillsWrapper>
                         <S.Description>Имею опыт работы со следующими технологиями:</S.Description>
                         <FlexWrapper wrap={'wrap'} align={'center'} justify={'space-around'}>
-                            {SkillData.map((s: SkillPropsType, index) => (
+                            {Data.SkillData.map((s: SkillPropsType, index) => (
                                 <Skill key={index} iconId={s.iconId} titleId={s.titleId}
                                        viewBox={s.viewBox ? s.viewBox : '0 0 48 48'}/>))}
                         </FlexWrapper>
